Split up util package into pkg/password, pkg/detach, pkg/version
Fixes: https://github.com/containers/common/issues/1654
