+++ libnetwork/cni/network.go
@@ -17,7 +17,7 @@ import (
 	"github.com/containernetworking/cni/libcni"
 	"github.com/containers/common/libnetwork/types"
 	"github.com/containers/common/pkg/config"
-	cutil "github.com/containers/common/pkg/util"
+	"github.com/containers/common/pkg/version"
 	"github.com/containers/storage/pkg/lockfile"
 	"github.com/containers/storage/pkg/unshare"
 	"github.com/sirupsen/logrus"
@@ -302,8 +302,8 @@ func (n *cniNetwork) NetworkInfo() types.NetworkInfo {
 	path := ""
 	packageVersion := ""
 	for _, p := range n.cniPluginDirs {
-		ver := cutil.PackageVersion(p)
-		if ver != cutil.UnknownPackage {
+		ver := version.Package(p)
+		if ver != version.UnknownPackage {
 			path = p
 			packageVersion = ver
 			break
@@ -317,8 +317,8 @@ func (n *cniNetwork) NetworkInfo() types.NetworkInfo {
 	}
 
 	dnsPath := filepath.Join(path, "dnsname")
-	dnsPackage := cutil.PackageVersion(dnsPath)
-	dnsProgram, err := cutil.ProgramVersionDnsname(dnsPath)
+	dnsPackage := version.Package(dnsPath)
+	dnsProgram, err := version.ProgramDnsname(dnsPath)
 	if err != nil {
 		logrus.Infof("Failed to get the dnsname plugin version: %v", err)
 	}
